pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lakeflow-jobs-meta"
version = "0.1.0"
description = "Metadata-driven framework for orchestrating Databricks Lakeflow Jobs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "databricks-sdk>=0.1.0",
    "pyspark>=3.0.0",
    "pyyaml>=6.0",
    "delta-spark>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
lakeflow-jobs-meta = "lakeflow_jobs_meta.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/lakeflow-jobs-meta"
Documentation = "https://github.com/yourusername/lakeflow-jobs-meta/docs"
Repository = "https://github.com/yourusername/lakeflow-jobs-meta"

